1st Semester/Sisteme avansate de codare si compresie multimedia/FinalProject/requirements.txt
numpy==1.19.4
Pillow==8.1.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyWavelets==1.1.1
